Appstore submission issue fixed

+++ FirstGame/utiliz.m
@@ -32,7 +32,7 @@
 {
     CGColorRef oldCGColor = oldColor.CGColor;
     
-    int numberOfComponents = CGColorGetNumberOfComponents(oldCGColor);
+    size_t numberOfComponents = CGColorGetNumberOfComponents(oldCGColor);
     // can not invert - the only component is the alpha
     if (numberOfComponents == 1) {
         return [UIColor colorWithCGColor:oldCGColor];
@@ -41,7 +41,7 @@
     const CGFloat *oldComponentColors = CGColorGetComponents(oldCGColor);
     CGFloat newComponentColors[numberOfComponents];
     
-    int i = numberOfComponents - 1;
+    int i = (int)numberOfComponents - 1;
     newComponentColors[i] = oldComponentColors[i]; // alpha
     while (--i >= 0) {
         newComponentColors[i] = 1 - oldComponentColors[i];
